pyproject.toml
[project]
name = "rag-v"
version = "0.1.0"
description = "RAG application using FAISS, Gemini LLM API, and LangGraph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.1.0",
    "langchain-google-genai>=0.1.0",
    "langchain-text-splitters>=0.1.0",
    "langgraph>=0.1.0",
    "faiss-cpu>=1.7.4",
    "google-generativeai>=0.3.0",
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "unstructured>=0.10.0",
    "pypdf>=3.15.0",
    "docx2txt>=0.8",
    "tiktoken>=0.5.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "pydantic>=2.4.2",
    # Voice capabilities
    "SpeechRecognition>=3.10.0",
    "gTTS>=2.3.2",
    "pydub>=0.25.1",
    "python-multipart>=0.0.6", # For handling file uploads in FastAPI
    "websockets>=11.0.3", # For real-time communication
    "pyttsx3>=2.98",
]
